// src/app/master/models/printformatm.ts
import { printformatd } from './printformatd';

import { GlobalVariables } from '../../core/models/globalvariables';

export class printformatm {

  blf_pkid: string;
  blf_type: string;
  blf_name: string;

  rec_mode: string;
  FormatList: printformatd[] = [];

  _globalvariables: GlobalVariables;

}