// src/app/master/models/printformatd.ts
import { GlobalVariables } from '../../core/models/globalvariables';
//EDIT-AJITH-06-10-2021

export class printformatd {

  blf_col_pkid: string;
  blf_col_name: string;
  blf_col_x: number;
  blf_col_y: number;
  blf_col_enabled: string;
  blf_col_height: number;
  blf_col_width: number;
  blf_col_font: string;
  blf_col_font_size: number;
  blf_col_style_font: string;
  blf_col_style_align: string;

  rec_mode: string;
  _globalvariables: GlobalVariables;

}